+++ src/Calendarien/calendar/Viewer/Viewer.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "./Viewer.css";
 
-import Week from "../Week";
 import DayOfWeek from "../DayOfWeek";
+import Week from "../Week";
 
 import { WEEK_COUNT, CREATE, START_DAY, LAST_DATE, MOVE_MONTH } from "../../utils/dates";
 import { line, createArray, ymd, pnCalculator } from "../../utils/util";